9624: Move filters components under Filters/
