+++ app/javascript/components/Filters/model.js
@@ -2,7 +2,7 @@ import isEmpty from 'lodash/isEmpty';
 import { action, observable, computed, reaction, toJS } from 'mobx';
 
 import ConditionSetModel from '../ConditionSetFormField/model';
-import { SUBMITTER_TYPES } from '../SubmitterFilter/component';
+import { SUBMITTER_TYPES } from './SubmitterFilter/component';
 
 /** Empty model to be used for resetting the store as needed. */
 const initialConditionSetData = Object.freeze(toJS(new ConditionSetModel({
